+++ src/api/func/java/org/smoothbuild/builtin/file/FilesFunction.java
@@ -6,7 +6,7 @@ import static org.smoothbuild.io.fs.base.RecursivePathsIterator.recursivePathsIt
 
 import java.io.IOException;
 
-import org.smoothbuild.exec.task.Container;
+import org.smoothbuild.exec.task.base.Container;
 import org.smoothbuild.io.fs.base.FileSystem;
 import org.smoothbuild.io.fs.base.Path;
 import org.smoothbuild.io.fs.base.PathIterator;