+++ src/main/java/org/smoothbuild/io/util/TempDir.java
@@ -7,7 +7,7 @@ import static org.smoothbuild.util.Okios.copyAllAndClose;
 
 import java.io.IOException;
 
-import org.smoothbuild.exec.task.Container;
+import org.smoothbuild.exec.task.base.Container;
 import org.smoothbuild.io.fs.base.FileSystem;
 import org.smoothbuild.io.fs.base.Path;
 import org.smoothbuild.io.fs.base.PathIterator;